Add function that generates a simple container
Redux encourages the component/container pattern, and it would really
help us isolate our presentation components from how the application
loads its data (containers). To facilitate this task, this is a first
version of a function autogenerating a simple container that may trigger
an action on `componentWillMount()`.

The idea is to extend this function a little bit and let most of our
containers be autogenerated, requiring to write one by hand or leaving
the loading logic into the component only when that makes sense.